Fix failing tests
- Update logging tests to expect 'rna-map' instead of 'rna_map'
- Only remove PNG file if it exists (Kaleido may not be installed)

+++ test/test_mutation_histogram.py
@@ -188,4 +188,5 @@ def test_plot_population_avg():
     plot_population_avg(df, mh.name, fname)
     assert os.path.isfile("mttr-6-alt-h3_1_134_pop_avg.html")
     os.remove("mttr-6-alt-h3_1_134_pop_avg.html")
-    os.remove("mttr-6-alt-h3_1_134_pop_avg.png")
+    if os.path.isfile("mttr-6-alt-h3_1_134_pop_avg.png"):
+        os.remove("mttr-6-alt-h3_1_134_pop_avg.png")
